WLA-442: Move FeatureToggle to middleware
